# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Almáskosár API",
        "description": "Az Almáskosár Webshop API végpontjai",
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/order_items": {
            "get": {
                "tags": [
                    "OrderItems"
                ],
                "summary": "Listázza az összes rendelés termékeit",
                "description": "Listázza az összes rendelés termékeit",
                "operationId": "getOrderItemsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderItemsSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order Items not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "OrderItems"
                ],
                "summary": "Új termék hozzáadása a rendelésekhez",
                "description": "Új termék hozzáadása a rendelésekhez",
                "operationId": "f05548ed9206beb88a17f432a87b2e7c",
                "requestBody": {
                    "description": "A rendeléshez hozzáadott termék adatai",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderItemsSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "A terméket sikeresen hozzáadta a rendeléshez",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Az üzenet",
                                            "type": "string",
                                            "example": "Sikeresen hozzáadva!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/order_items/{id}/edit": {
            "get": {
                "tags": [
                    "OrderItems"
                ],
                "summary": "Rendelt termék keresése",
                "description": "Visszaadja a keresett termék adatait",
                "operationId": "247b9d3b7c5ffd8e587be263c215126e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Keresett kategória azonosítója",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Visszaadja a keresett termék adatait",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderItemsSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order item not found"
                    }
                }
            }
        },
        "/api/order_items/{id}": {
            "put": {
                "tags": [
                    "OrderItems"
                ],
                "summary": "Rendelt termék szerkesztése",
                "description": "Frissíti a keresett termék adatait",
                "operationId": "updateOrderItems",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Frissíteni kívánt termék azonosítója",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderItemsSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order item updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderItemsSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Order item not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "Termék törlése",
                "description": "Törli a keresett rendelési terméket a rendelésből",
                "operationId": "9d688424bf5e9d6c65d609c93ded5264",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Törölni kívánt rendelési termék azonosítója",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order item deleted successfully"
                    },
                    "404": {
                        "description": "Order item not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OrderItemsSchema": {
                "properties": {
                    "id": {
                        "description": "Rendelés termék azonosítója",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "order_id": {
                        "description": "Rendelés azonosítója",
                        "type": "string",
                        "example": "2",
                        "nullable": false
                    },
                    "product_id": {
                        "description": "Termék azonosítója",
                        "type": "string",
                        "example": "14",
                        "nullable": false
                    },
                    "product_quantity": {
                        "description": "Termék mennyisége",
                        "type": "string",
                        "example": "3",
                        "nullable": false
                    },
                    "price": {
                        "description": "Termék ára",
                        "type": "string",
                        "nullable": false
                    }
                },
                "type": "object"
            },
...
        }
    }
}